Develop

+++ TaskManager/DataModels/DataModel.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using LinqToDB;
 using TaskManager.DataModelComponents;
 using TaskManager.Components;
+using System.Linq.Expressions;
 
 namespace TaskManager.DataModels
 {
@@ -26,6 +27,9 @@ namespace TaskManager.DataModels
 
         public void AddTask(UserTask task)
         {
+            if (task == null)
+                throw new ArgumentNullException(nameof(task));
+
             using (var db = new UserTasksDB())
                 db.Insert(task);
 
@@ -34,6 +38,9 @@ namespace TaskManager.DataModels
 
         public void UpdateTask(UserTask task)
         {
+            if (task == null)
+                throw new ArgumentNullException(nameof(task));
+
             using (var db = new UserTasksDB())
                 db.Update(task);
 
@@ -42,6 +49,12 @@ namespace TaskManager.DataModels
 
         public void DeleteTask(UserTask task)
         {
+            if (task == null)
+                throw new ArgumentNullException(nameof(task));
+
+            if (task.Id <= 0) //С 1?
+                throw new ArgumentOutOfRangeException($"Id of {task} should be positive");
+
             using (var db = new UserTasksDB())
                 db.Delete(task);
 
@@ -63,6 +76,9 @@ namespace TaskManager.DataModels
 
         public List<UserTask> GetTasksOfDay(string date)
         {
+            if (string.IsNullOrWhiteSpace(date))
+                throw new ArgumentException($"{nameof(date)} should not to be null or empty.");
+
             List<UserTask> tasks;
             using (var db = new UserTasksDB())
             {
@@ -72,6 +88,37 @@ namespace TaskManager.DataModels
                 tasks = filterResult.ToList();
             }
             return tasks;
+        }
+
+        public List<UserTask> GetTasksOfDays(string beginDate, string endDate)
+        {
+            if (string.IsNullOrWhiteSpace(beginDate) || string.IsNullOrWhiteSpace(endDate))
+                throw new ArgumentException($"{nameof(beginDate)} or {nameof(endDate)} should not to be null or empty.");
+
+            List<UserTask> tasks;
+            Expression<Func<UserTask, bool>> compare;
+
+            if (beginDate == endDate) //Вынести в поле класса?
+                compare = t => t.TaskDate == beginDate;
+            else
+                compare = t => t.TaskDate.CompareTo(beginDate) >= 0 && t.TaskDate.CompareTo(endDate) <= 0;
+
+            using (var db = new UserTasksDB())
+            {
+                var query = db.UserTasks.Where(compare);
+                var filterResult = _taskFilter.Filter(query, _filterType);
+
+                tasks = filterResult.ToList();
+            }
+            return tasks;
+        }
+
+        public List<string> GetAllTaskDates()
+        {
+            using (var db = new UserTasksDB())
+            {
+                return db.UserTasks.Select(t => t.TaskDate).ToList();
+            }
         }
 
         public void SortBy(SortType sort)
